requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
APScheduler==3.8.0
atomicwrites==1.4.0
attrs==19.3.0
BeautifulReport==0.1.2
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2020.4.5.2
chainmap==1.0.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
combomethod==1.0.12
configparser==5.0.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
et-xmlfile==1.0.1
fake-useragent==0.1.11
Flask==2.0.2
idna==2.9
importlib-metadata==1.6.1
iniconfig==1.1.1
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.2
jsonpath==0.82
jsonpath-rw==1.4.0
kiwisolver==1.3.1
Logbook==1.5.3
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.3.4
more-itertools==8.4.0
nulltype==2.3.1
numpy==1.19.5
openpyxl==3.0.4
options==1.4.10
packaging==20.4
pact-python==1.2.0
pactverify==1.7.5
ParamUnittest==0.2
Pillow==8.3.1
pluggy==0.13.1
ply==3.11
psutil==5.7.2
py==1.8.1
pygame==2.0.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.11.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.9.1
selenium==3.14.0
six==1.12.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.25.9
wcwidth==0.2.4
Werkzeug==2.0.2
win32compat==221.26
win32core==221.36
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.1.0
